Update manage-data-retention.mdx
Offering this edit to "copy retention" section since it was brought to my attention that this feature is only for copying retention from child-to-child. By default retention is inherited from the parent. If a child account retention is updated, this breaks inheritance. If your other child-accounts also need inheritance broken from the parent, this feature is available to easily edit those other child-accounts.